move tihngs around and add all join and create room is safe, next step is to start new round
