+++ front/src/Components/home/home.component.ts
@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 import { Router } from '@angular/router';
+;
+import { MatButtonModule } from '@angular/material/button';
 import { JoinRoomComponent } from '../join-room/join-room.component';
 import { CreateRoomComponent } from '../create-room/create-room.component';
-import { MatButtonModule } from '@angular/material/button';
 
 @Component({
   selector: 'app-home',
