+++ front/src/Components/join-room/join-room.component.css
@@ -1,0 +1,21 @@
+.join-room {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+  min-width: fit-content;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin-top: 20px;
+  min-width: fit-content;
+}
+
+button {
+  margin: 10px;
+}
