+++ front/src/Components/room/room.component.html
@@ -1,0 +1,40 @@
+<div class="room-container">
+  <mat-card>
+    <mat-card-header>
+      <mat-card-title>Room Code: {{ roomCode }}</mat-card-title>
+    </mat-card-header>
+    <mat-card-content>
+      <h3>Players</h3>
+      <mat-list>
+        <mat-list-item *ngFor="let player of players">
+          {{ player.username }} ({{ player.language }})
+        </mat-list-item>
+      </mat-list>
+
+      <h3 *ngIf="!currentRound">Rounds Not Started</h3>
+      <button
+        mat-raised-button
+        color="primary"
+        *ngIf="!currentRound"
+        (click)="startRounds()"
+      >
+        Start Rounds
+      </button>
+
+      <mat-tab-group *ngIf="currentRound">
+        <mat-tab
+          *ngFor="let round of rounds; let i = index"
+          label="Round {{ i + 1 }}"
+        >
+          <p>Round {{ i + 1 }} Details: {{ round.details }}</p>
+        </mat-tab>
+      </mat-tab-group>
+
+      <mat-progress-bar
+        *ngIf="currentRound"
+        mode="determinate"
+        [value]="progressValue"
+      ></mat-progress-bar>
+    </mat-card-content>
+  </mat-card>
+</div>
