+++ front/src/app/app.component.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { HomeComponent } from './home/home.component';
-import { CreateRoomComponent } from './create-room/create-room.component';
-import { JoinRoomComponent } from './join-room/join-room.component';
-import { RoomComponent } from './room/room.component';
+import { HomeComponent } from '../Components/home/home.component';
+import { CreateRoomComponent } from '../Components/create-room/create-room.component';
+import { JoinRoomComponent } from '../Components/join-room/join-room.component';
+import { RoomComponent } from '../Components/room/room.component';
 
 @Component({
   selector: 'app-root',
