+++ front/src/app/app.routes.ts
@@ -1,8 +1,8 @@
 import { Routes } from '@angular/router';
-import { HomeComponent } from './home/home.component';
-import { CreateRoomComponent } from './create-room/create-room.component';
-import { JoinRoomComponent } from './join-room/join-room.component';
-import { RoomComponent } from './room/room.component';
+import { HomeComponent } from '../Components/home/home.component';
+import { CreateRoomComponent } from '../Components/create-room/create-room.component';
+import { JoinRoomComponent } from '../Components/join-room/join-room.component';
+import { RoomComponent } from '../Components/room/room.component';
 
 export const routes: Routes = [
   { path: '', component: HomeComponent },
