--- front/src/app/create-room/create-room.component.html
@@ -1,28 +1,0 @@
-<div class="create-room">
-  <h2>Create Room</h2>
-  <form [formGroup]="createRoomForm" (submit)="createRoom()">
-    <mat-form-field>
-      <mat-label>Username</mat-label>
-      <input matInput formControlName="username" />
-    </mat-form-field>
-    <mat-form-field>
-      <mat-label>Choose a language</mat-label>
-      <input
-        matInput
-        formControlName="language"
-        required
-        (blur)="updateErrorMessage()"
-        [matAutocomplete]="auto"
-      />
-      <mat-autocomplete #auto="matAutocomplete">
-        @for (option of filteredLanguages | async; track option) {
-          <mat-option [value]="option">{{ option }}</mat-option>
-        }
-      </mat-autocomplete>
-      @if (errorMessage()) {
-        <mat-error>{{ errorMessage() }}</mat-error>
-      }
-    </mat-form-field>
-    <button mat-fab extended  [disabled]="createRoomForm.invalid">Create</button>
-  </form>
-</div>
