--- front/src/app/join-room/join-room.component.html
@@ -1,9 +1,0 @@
-<div class="join-room">
-  <h2>Join Room</h2>
-  <form (submit)="joinRoom()" [formGroup]="joinRoomForm">
-    <input formControlName="roomCode" placeholder="Room Code" />
-    <input formControlName="username" placeholder="Username" />
-    <input formControlName="language" placeholder="Language" />
-    <button (click)="joinRoom()" class="primary">Join Room</button>
-  </form>
-</div>
