--- front/src/app/room/room.component.html
@@ -1,9 +1,0 @@
-<p>Room cretaed: {{ this.roomCode }}</p>
-
-<div>
-  <div *ngFor="let player of players">
-    <span>{{ player.username }}</span>
-    <span> - </span>
-    <span>{{ player.language }}</span>
-  </div>
-</div>
